Fix race condition
Our first fix solved the issue of the pixels subscribing before the iframe was loaded. Since the iframe was not loaded, the events where being triggered before the JS of the iframe was listening to events, making the iframe missing those events.

Now, the issue is: the events were being triggered *before* the pixel could even start subscribing. This made those events to be lost as well.

The current solution keeps all the first 7 seconds of triggered events in a variable that could be retrieved by the PixelIFrame once it loads.